cve bug fix for breaking page
