+++ src/components/security/AddCveModal.tsx
@@ -18,37 +18,9 @@ import React, { Component } from 'react'
 import { showError, Progressing, VisibleModal, CustomInput } from '@devtron-labs/devtron-fe-common-lib'
 import { getCVEPolicies } from './security.service'
 import { CVE_ID_NOT_FOUND, ViewType } from '../../config'
-import { VulnerabilityAction } from './security.types'
+import { AddCveModalProps, AddCveModalState, VulnerabilityAction } from './security.types'
 import { ReactComponent as Close } from '../../assets/icons/ic-close.svg'
 import { ReactComponent as Info } from '../../assets/icons/ic-info-outline.svg'
-
-export interface AddCveModalProps {
-    close: () => void
-    saveCVE: (cve: string, policy: VulnerabilityAction) => void
-}
-
-export interface AddCveModalState {
-    view: string
-    cve: string
-    cveError: boolean
-    policy: VulnerabilityAction
-    clusters: {
-        name: string
-        policy: VulnerabilityAction
-        isCollapsed: boolean
-        environments: [
-            {
-                name: string
-                isCollapsed: boolean
-                policy: VulnerabilityAction
-                applications: {
-                    name: string
-                    policy: VulnerabilityAction
-                }[]
-            },
-        ]
-    }[]
-}
 
 export class AddCveModal extends Component<AddCveModalProps, AddCveModalState> {
     constructor(props) {
@@ -57,7 +29,6 @@ export class AddCveModal extends Component<AddCveModalProps, AddCveModalState> {
             view: ViewType.FORM,
             cve: '',
             policy: VulnerabilityAction.allow,
-            cveError: false,
             clusters: [],
         }
         this.handleCveChange = this.handleCveChange.bind(this)
@@ -79,7 +50,7 @@ export class AddCveModal extends Component<AddCveModalProps, AddCveModalState> {
         if (regex.test(cve)) {
             this.props.saveCVE(this.state.cve, this.state.policy)
         } else {
-            this.setState({ cveError: true })
+            this.props.setCVEError()
         }
     }
 
@@ -94,7 +65,8 @@ export class AddCveModal extends Component<AddCveModalProps, AddCveModalState> {
                 })
             })
             .catch((error) => {
-                showError(error)
+                this.props.setCVEError()
+                this.setState({ view: ViewType.FORM })
             })
     }
 
@@ -197,7 +169,7 @@ export class AddCveModal extends Component<AddCveModalProps, AddCveModalState> {
                                     value={this.state.cve}
                                     onChange={this.handleCveChange}
                                     isRequiredField
-                                    error={this.state.cveError && CVE_ID_NOT_FOUND}
+                                    error={this.props.cveError && CVE_ID_NOT_FOUND}
                                 />
                             </label>
                             {/* <button type="submit" className="cta mb-5" tabIndex={2} onClick={this.searchCVE}>Search</button> */}
@@ -211,7 +183,7 @@ export class AddCveModal extends Component<AddCveModalProps, AddCveModalState> {
                                     tabIndex={1}
                                     onClick={this.handlePolicyChange}
                                     checked={this.state.policy === VulnerabilityAction.allow}
-                                />{' '}
+                                />&nbsp;
                                 <span className="ml-10 mr-5">Allow</span>
                             </label>
                             <label className="form__label form__label--flex cursor ml-10">
@@ -237,9 +209,7 @@ export class AddCveModal extends Component<AddCveModalProps, AddCveModalState> {
                                 <span className="ml-10 mr-5">Block if fix is available</span>
                             </label>
                         </div>
-                        {/* <hr className="mt-10 mb-20"></hr>
-                    {this.renderList()}
-                    <hr className="mt-6 mb-16"></hr> */}
+                     
                         <div className="flex right form-row">
                             <button
                                 type="button"
