+++ src/components/security/SecurityPolicyEdit.tsx
@@ -35,6 +35,7 @@ import {
     VulnerabilityAction,
     ResourceLevel,
     VulnerabilityPolicy,
+    SecurityPolicyEditState,
 } from './security.types'
 import { AddCveModal } from './AddCveModal'
 import { ReactComponent as Arrow } from '../../assets/icons/ic-chevron-down.svg'
@@ -46,7 +47,7 @@ import { getCustomOptionSelectionStyle } from '../v2/common/ReactSelect.utils'
 
 export class SecurityPolicyEdit extends Component<
     FetchPolicyQueryParams,
-    GetVulnerabilityPolicyResponse & { showWhitelistModal: boolean; view: string }
+    GetVulnerabilityPolicyResponse & SecurityPolicyEditState
 > {
     private vulnerabilityMetaData: VulnerabilityUIMetaData[] = [
         {
@@ -103,6 +104,7 @@ export class SecurityPolicyEdit extends Component<
                 level: this.props.level,
                 policies: [],
             },
+            cveError: false,
         }
         this.toggleAddCveModal = this.toggleAddCveModal.bind(this)
         this.updateSeverity = this.updateSeverity.bind(this)
@@ -139,8 +141,12 @@ export class SecurityPolicyEdit extends Component<
                 }
             })
             .catch((error) => {
-                showError(error)
-                this.setState({ view: ViewType.ERROR })
+                if(error.code === 404){
+                    this.setState({ cveError: true })
+                } else {
+                    showError(error)
+                    this.setState({ view: ViewType.ERROR })
+                }
             })
     }
 
@@ -558,6 +564,10 @@ export class SecurityPolicyEdit extends Component<
         }
 
         const isCollapsible = this.props.level === 'application'
+
+        const setCVEErrorToTrue = () => {
+                this.setState({ cveError: true })
+        }
         return (
             <>
                 {this.renderHeader()}
@@ -592,7 +602,7 @@ export class SecurityPolicyEdit extends Component<
                     )
                 })}
                 {this.state.showWhitelistModal ? (
-                    <AddCveModal saveCVE={this.saveCVE} close={this.toggleAddCveModal} />
+                    <AddCveModal saveCVE={this.saveCVE} close={this.toggleAddCveModal} cveError={this.state.cveError} setCVEError={setCVEErrorToTrue} />
                 ) : null}
             </>
         )
